Reorg p2p message handlers
